feat: add priority map to preprocess kernel and centerpoint_trt
This commit adds the `priority_map` array to the `preprocess_kernel.hpp` and `centerpoint_trt.hpp` files. The `priority_map` is a new CUDA array that is allocated and initialized in the `PreprocessKernelTest::SetUp()` function. The `priority_map` is used in the `generateBaseFeatures_launch()` function to generate base features for the voxel grid. The `priority_map` is also used in the `centerpoint_trt.hpp` file for initializing and updating the `priority_map_` member variable.

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>